+++ BetterKick/CHANGELOG.md
@@ -1,3 +1,12 @@
+# v0.1.1 - Stop cheaters
+
+- **EVERYONE** must have the mod installed to kick.
+- You can no longer kick when climbing.
+- You can no longer kick when on 0 stamina
+- You can no longer kick when typing using a text chat mod.
+- Added GUI for when you cannot kick because people do not have the mod installed.
+- Delisted version 0.1.0
+
 # v0.1.0
 
 - Mod Release